// client/src/app/models/user.ts
export class User {
    id: number;
    name: string;
    description: string;
    groups:string[];
    username: string;
    email: string;
    address:string;
    password:string;
    confirm_password:string;
    phone_no: string;
    active: boolean;
    created_at: string;
    updated_at: string;
}